+++ sdk/telemetry/test/doc.go
@@ -1,9 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-// Package test provides isolated testing for
-// go.opentelemetry.io/auto/sdk/telemetry. It exists to ensure that package is
-// tested without importing any OTLP or pdata dependencies.
-//
-// Deprecated: This module is no longer supported.
-package test